project/requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
flask-wtf==1.1.1
paho-mqtt==1.6.1
python-dotenv==1.0.0
requests==2.28.2
gunicorn==20.1.0
werkzeug==2.2.3
jinja2==3.1.2
SQLAlchemy==2.0.4
alembic==1.10.2
itsdangerous==2.1.2
click==8.1.3
markupsafe==2.1.2
wtforms==3.0.1 